Split multimodal group into vision and audio
